requirements.txt
streamlit==1.32.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.2
python-jose==3.3.0
reportlab==4.0.4
requests==2.31.0
PyPDF2==3.0.1
python-multipart==0.0.9
pydantic==2.6.1
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
passlib==1.7.4 